[UPD] generalised notes list view
