+++ cogs/moderation/whois.py
@@ -1,11 +1,8 @@
 import nextcord
 from nextcord.ext import commands
-from utils.base import PageView
 
 from utils.config import Configuration
-from utils.notes import fetch_notes
 from utils.perms import is_mod, permcheck
-from utils.sersi_embed import SersiEmbed
 from utils.whois import create_whois_embed, WhoisView
 
 
@@ -110,41 +107,6 @@ class WhoisSystem(commands.Cog):
                 embed=await create_whois_embed(self.config, interaction, user),
                 view=WhoisView(user.id),
             )
-
-    @commands.Cog.listener()
-    async def on_interaction(self, interaction: nextcord.Interaction):
-        try:
-            btn_id = interaction.data["custom_id"]
-        except KeyError:
-            return
-
-        if not btn_id.startswith("whois-"):
-            return
-
-        await interaction.response.defer(ephemeral=True)
-
-        match btn_id.split(":", 1):
-            case ["whois-notes", user_id]:
-                note_embed = SersiEmbed(title=f"{interaction.guild.name} Notes")
-                note_embed.set_thumbnail(interaction.guild.icon.url)
-
-                view = PageView(
-                    config=self.config,
-                    base_embed=note_embed,
-                    fetch_function=fetch_notes,
-                    author=interaction.user,
-                    entry_form="{entry}",
-                    field_title="{entries[0].list_entry_header}",
-                    inline_fields=False,
-                    cols=5,
-                    per_col=1,
-                    init_page=1,
-                    ephemeral=True,
-                    member_id=user_id,
-                    author_id=None,
-                )
-
-                await view.send_followup(interaction)
 
 
 def setup(bot: commands.Bot, **kwargs):
